[Components] [Post] Moved post functionality to its own component.
